refactor: move upload analytics processor
