--- analytics/upload_processor.py
@@ -1,3 +1,0 @@
-from services.upload_processing import UploadAnalyticsProcessor
-
-__all__ = ["UploadAnalyticsProcessor"]
