+++ services/registry.py
@@ -47,7 +47,9 @@ get_service = registry.get_service
 register_service("FileProcessor", "services.data_processing.unified_file_validator:UnifiedFileValidator")
 register_service("FileHandler", "services.data_processing.unified_file_validator:UnifiedFileValidator")
 register_service("UnifiedFileValidator", "services.data_processing.unified_file_validator:UnifiedFileValidator")
-register_service("UploadAnalyticsProcessor", "services.upload_processing:UploadAnalyticsProcessor")
+register_service(
+    "UploadAnalyticsProcessor", "services.analytics.upload_analytics:UploadAnalyticsProcessor"
+)
 
 register_service(
     "get_analytics_service", "services.analytics_service:get_analytics_service"
