+++ tests/test_process_and_analyze.py
@@ -2,7 +2,7 @@ import pandas as pd
 
 from services.data_processing.file_processor import FileProcessor
 
-from analytics.upload_processor import UploadAnalyticsProcessor
+from services.analytics.upload_analytics import UploadAnalyticsProcessor
 from services.file_processing_service import FileProcessingService
 from services.data_validation import DataValidationService
 from services.data_processing.processor import Processor
